requirements.txt
python-binance==1.0.12
flask==2.0.1